[BACKEND] Services divided to implementation and interface.
Divide AccountService to some smaller services.
