+++ app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/utils/services/MailSenderService.java
@@ -1,0 +1,7 @@
+package pl.lodz.p.it.inz.sgruda.multiStore.utils.services;
+
+import javax.mail.MessagingException;
+
+public interface MailSenderService {
+    void sendMail(String to, String subject, String text, boolean isHtmlContent) throws MessagingException;
+}
