+++ app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/utils/services/MailSenderServiceImpl.java
@@ -1,0 +1,31 @@
+package pl.lodz.p.it.inz.sgruda.multiStore.utils.services;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.MimeMessageHelper;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Service;
+
+import javax.mail.MessagingException;
+import javax.mail.internet.MimeMessage;
+
+@Service
+public class MailSenderServiceImpl implements MailSenderService {
+    private JavaMailSender javaMailSender;
+
+    @Autowired
+    public MailSenderServiceImpl(JavaMailSender javaMailSender) {
+        this.javaMailSender = javaMailSender;
+    }
+
+    @Async
+    @Override
+    public void sendMail(String to, String subject, String text, boolean isHtmlContent) throws MessagingException {
+        MimeMessage mimeMessage = javaMailSender.createMimeMessage();
+        MimeMessageHelper mimeMessageHelper = new MimeMessageHelper(mimeMessage, true);
+        mimeMessageHelper.setTo(to);
+        mimeMessageHelper.setSubject(subject);
+        mimeMessageHelper.setText(text, isHtmlContent);
+        javaMailSender.send(mimeMessage);
+    }
+}
